requirements.txt
pyqlib>=0.9.3
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
torch>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0
loguru>=0.7.0
tabulate>=0.9.0
openai>=1.0.0
anthropic>=0.18.0
tenacity>=8.0.0
jieba>=0.42.1
snownlp>=0.12.3
